requirements.txt
llama-cpp-python==0.3.1
gradio==4.44.0